requirements.txt
requests
jinja2
python-dotenv
google-genai
xai-sdk
